refactor(meanings): refactors meaning retriever
